Moved Series.drawGraph to LineSeries.drawGraph.
