+++ js/Series/Line/LineSeries.js
@@ -25,7 +25,7 @@ import palette from '../../Core/Color/Palette.js';
 import Series from '../../Core/Series/Series.js';
 import SeriesRegistry from '../../Core/Series/SeriesRegistry.js';
 import U from '../../Core/Utilities.js';
-var merge = U.merge;
+var defined = U.defined, merge = U.merge;
 /* *
  *
  *  Class
@@ -143,6 +143,138 @@ var LineSeries = /** @class */ (function (_super) {
                 graph.isArea = graphPath.isArea; // For arearange animation
             }
         });
+    };
+    // eslint-disable-next-line valid-jsdoc
+    /**
+     * Get the graph path.
+     *
+     * @private
+     */
+    LineSeries.prototype.getGraphPath = function (points, nullsAsZeroes, connectCliffs) {
+        var series = this, options = series.options, step = options.step, reversed, graphPath = [], xMap = [], gap;
+        points = points || series.points;
+        // Bottom of a stack is reversed
+        reversed = points.reversed;
+        if (reversed) {
+            points.reverse();
+        }
+        // Reverse the steps (#5004)
+        step = {
+            right: 1,
+            center: 2
+        }[step] || (step && 3);
+        if (step && reversed) {
+            step = 4 - step;
+        }
+        // Remove invalid points, especially in spline (#5015)
+        points = this.getValidPoints(points, false, !(options.connectNulls && !nullsAsZeroes && !connectCliffs));
+        // Build the line
+        points.forEach(function (point, i) {
+            var plotX = point.plotX, plotY = point.plotY, lastPoint = points[i - 1], 
+            // the path to this point from the previous
+            pathToPoint;
+            if ((point.leftCliff || (lastPoint && lastPoint.rightCliff)) &&
+                !connectCliffs) {
+                gap = true; // ... and continue
+            }
+            // Line series, nullsAsZeroes is not handled
+            if (point.isNull && !defined(nullsAsZeroes) && i > 0) {
+                gap = !options.connectNulls;
+                // Area series, nullsAsZeroes is set
+            }
+            else if (point.isNull && !nullsAsZeroes) {
+                gap = true;
+            }
+            else {
+                if (i === 0 || gap) {
+                    pathToPoint = [[
+                            'M',
+                            point.plotX,
+                            point.plotY
+                        ]];
+                    // Generate the spline as defined in the SplineSeries object
+                }
+                else if (series.getPointSpline) {
+                    pathToPoint = [series.getPointSpline(points, point, i)];
+                }
+                else if (step) {
+                    if (step === 1) { // right
+                        pathToPoint = [[
+                                'L',
+                                lastPoint.plotX,
+                                plotY
+                            ]];
+                    }
+                    else if (step === 2) { // center
+                        pathToPoint = [[
+                                'L',
+                                (lastPoint.plotX + plotX) / 2,
+                                lastPoint.plotY
+                            ], [
+                                'L',
+                                (lastPoint.plotX + plotX) / 2,
+                                plotY
+                            ]];
+                    }
+                    else {
+                        pathToPoint = [[
+                                'L',
+                                plotX,
+                                lastPoint.plotY
+                            ]];
+                    }
+                    pathToPoint.push([
+                        'L',
+                        plotX,
+                        plotY
+                    ]);
+                }
+                else {
+                    // normal line to next point
+                    pathToPoint = [[
+                            'L',
+                            plotX,
+                            plotY
+                        ]];
+                }
+                // Prepare for animation. When step is enabled, there are
+                // two path nodes for each x value.
+                xMap.push(point.x);
+                if (step) {
+                    xMap.push(point.x);
+                    if (step === 2) { // step = center (#8073)
+                        xMap.push(point.x);
+                    }
+                }
+                graphPath.push.apply(graphPath, pathToPoint);
+                gap = false;
+            }
+        });
+        graphPath.xMap = xMap;
+        series.graphPath = graphPath;
+        return graphPath;
+    };
+    // eslint-disable-next-line valid-jsdoc
+    /**
+     * Get zones properties for building graphs. Extendable by series with
+     * multiple lines within one series.
+     *
+     * @private
+     */
+    LineSeries.prototype.getZonesGraphs = function (props) {
+        // Add the zone properties if any
+        this.zones.forEach(function (zone, i) {
+            var propset = [
+                'zone-graph-' + i,
+                'highcharts-graph highcharts-zone-graph-' + i + ' ' +
+                    (zone.className || '')
+            ];
+            if (!this.chart.styledMode) {
+                propset.push((zone.color || this.color), (zone.dashStyle || this.options.dashStyle));
+            }
+            props.push(propset);
+        }, this);
+        return props;
     };
     /**
      * General options for all series types.
