+++ js/Series/Polygon/PolygonSeries.js
@@ -25,7 +25,7 @@ import H from '../../Core/Globals.js';
 var noop = H.noop;
 import LegendSymbolMixin from '../../Mixins/LegendSymbol.js';
 import SeriesRegistry from '../../Core/Series/SeriesRegistry.js';
-var Series = SeriesRegistry.series, _a = SeriesRegistry.seriesTypes, AreaSeries = _a.area, ScatterSeries = _a.scatter;
+var Series = SeriesRegistry.series, _a = SeriesRegistry.seriesTypes, AreaSeries = _a.area, LineSeries = _a.line, ScatterSeries = _a.scatter;
 import U from '../../Core/Utilities.js';
 var extend = U.extend, merge = U.merge;
 import '../../Core/Legend.js';
@@ -54,7 +54,7 @@ var PolygonSeries = /** @class */ (function (_super) {
      *
      * */
     PolygonSeries.prototype.getGraphPath = function () {
-        var graphPath = Series.prototype.getGraphPath.call(this), i = graphPath.length + 1;
+        var graphPath = LineSeries.prototype.getGraphPath.call(this), i = graphPath.length + 1;
         // Close all segments
         while (i--) {
             if ((i === graphPath.length || graphPath[i][0] === 'M') && i > 0) {
