+++ js/Series/Waterfall/WaterfallSeries.js
@@ -612,6 +612,7 @@ var WaterfallSeries = /** @class */ (function (_super) {
     return WaterfallSeries;
 }(ColumnSeries));
 extend(WaterfallSeries.prototype, {
+    getZonesGraphs: LineSeries.prototype.getZonesGraphs,
     pointValKey: 'y',
     // Property needed to prevent lines between the columns from disappearing
     // when negativeColor is used.
