+++ ts/Series/Waterfall/WaterfallSeries.ts
@@ -932,12 +932,14 @@ class WaterfallSeries extends ColumnSeries {
 }
 
 interface WaterfallSeries {
+    getZonesGraph: typeof LineSeries.prototype.getZonesGraphs;
     pointClass: typeof WaterfallPoint;
     pointValKey: string;
     showLine: boolean;
 }
 
 extend(WaterfallSeries.prototype, {
+    getZonesGraphs: LineSeries.prototype.getZonesGraphs,
     pointValKey: 'y',
     // Property needed to prevent lines between the columns from disappearing
     // when negativeColor is used.
